update refresh colors button
