+++ app/src/components/ImageContainerHeader.tsx
@@ -1,7 +1,9 @@
-import React from "react";
+import React, { useState } from "react";
 import styled from "styled-components";
 import { useRecoilState, useRecoilValue } from "recoil";
 import { Checkbox } from "@material-ui/core";
+import { Autorenew } from "@material-ui/icons";
+import { animated, useSpring } from "react-spring";
 
 import DropdownHandle from "./DropdownHandle";
 import SelectionMenu from "./SelectionMenu";
@@ -28,11 +30,9 @@ const Wrapper = styled.div`
 const SamplesHeader = styled.div`
   display: flex;
   justify-content: space-between;
-  height: 45px;
   overflow-x: hidden;
   margin-left: 1rem;
   margin-right: -1rem;
-  padding: 0.5rem 0;
   flex-grow: 1;
 `;
 
@@ -43,15 +43,63 @@ const OptionsContainer = styled.div`
 const OptionContainer = styled.div`
   display: flex;
   justify-content: space-between;
-  background: ${({ theme }) => theme.backgroundDark};
-  box-shadow: 0 8px 15px 0 rgba(0, 0, 0, 0.43);
-  border: 1px solid #191c1f;
-  border-radius: 2px;
+  border-right: 1px solid #191c1f;
   color: ${({ theme }) => theme.fontDark};
-  margin-top: 0.25rem;
   font-weight: bold;
   cursor: pointer;
 `;
+
+const Button = animated(styled.div`
+  cursor: pointer;
+  width: 100%;
+  margin-top: 3px;
+  margin-left: 0;
+  margin-right: 0;
+  padding: 0 0.2em;
+  border-radius: 2px;
+  display: flex;
+  height: 32px;
+`);
+
+const ButtonText = styled.div`
+  padding-right: 4px;
+  padding-left: 2px;
+  white-space: nowrap;
+  overflow-x: hidden;
+  text-overflow: ellipsis;
+  font-weight: bold;
+  padding-top: 4px;
+  letter-spacing: 0.00938em;
+  line-height: 24px;
+`;
+
+const RefreshButton = () => {
+  const theme = useTheme();
+  const [colorSeed, setColorSeed] = useRecoilState(atoms.colorSeed);
+  const [clicked, setClicked] = useState(false);
+  const props = useSpring({
+    backgroundColor: clicked ? theme.backgroundLight : theme.background,
+    color: clicked ? theme.font : theme.fontDark,
+    onRest: () => clicked && setClicked(false),
+    config: {
+      duration: 250,
+    },
+  });
+  return (
+    <Button
+      style={props}
+      onClick={() => {
+        setColorSeed(colorSeed + 1);
+        setClicked(true);
+      }}
+    >
+      <div style={{ marginTop: 4 }}>
+        <Autorenew />
+      </div>
+      <ButtonText>Refresh field colors</ButtonText>
+    </Button>
+  );
+};
 
 const ImageContainerHeader = ({ showSidebar, onShowSidebar }: Props) => {
   const totalCount = useRecoilValue(selectors.totalCount);
@@ -78,13 +126,24 @@ const ImageContainerHeader = ({ showSidebar, onShowSidebar }: Props) => {
       />
       <SamplesHeader>
         <OptionsContainer>
-          <SelectionMenu />
           <OptionContainer onClick={() => setColorByLabel(!colorByLabel)}>
-            <span style={{ height: "2rem", padding: "0.5rem" }}>
+            <span
+              style={{
+                height: "2rem",
+                padding: "0.6rem 0",
+                pointerEvents: "none",
+              }}
+            >
               Color by label
             </span>
             <Checkbox style={{ color: theme.brand }} checked={colorByLabel} />
           </OptionContainer>
+          <OptionContainer onClick={() => setColorByLabel(!colorByLabel)}>
+            <RefreshButton />
+          </OptionContainer>
+          <div>
+            <SelectionMenu />
+          </div>
         </OptionsContainer>
         {countStr !== null ? (
           <div className="total" style={{ paddingRight: "1rem" }}>
